add SendMessageWriteDatabaseProcessor

+++ imsdk/src/main/java/com/masonsoft/imsdk/IMMessage.java
@@ -125,7 +125,7 @@ public class IMMessage {
      * 消息的发送状态
      */
     @NonNull
-    public final StateProp<Long> sendState = new StateProp<>();
+    public final StateProp<Integer> sendState = new StateProp<>();
 
     /**
      * 消息的发送进度. 范围 [0, 1]
